[naga] Move methods on TypeInner and friends into their own module.
The `POINTER_SPAN` constant was not used outside the module, so make
it private.

No other code changes.
